+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v1.0.0
 - Fix inline variable names in the `*c` macros
-- New macros: `eprintc`, `eprintcln`
+- New formatting macros: `eprintc`, `eprintcln`
+- Unformatting macros: `formatnc`, `printnc`, `printncln`, `eprintnc`,
+  `eprintncln`
 
 ## v0.1.0
 The first version